+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/), and follows [
 
 ## [Unreleased]
 ### Added
+- Select photos by date in Added, Modified and Accessed views
+
 ### Changed
 ### Deprecated
 ### Removed